BLUEZ-6: Fix CSRF token not persisting in session storage

+++ src/Controller/OAuthController.php
@@ -4,6 +4,8 @@ namespace Drupal\zendesk_connect\Controller;
 
 use Drupal\Component\Utility\Crypt;
 use Drupal\Core\Controller\ControllerBase;
+use Drupal\Core\Routing\TrustedRedirectResponse;
+use Drupal\Core\Session\SessionManagerInterface;
 use Drupal\user\Entity\User;
 use Drupal\user\PrivateTempStore;
 use League\OAuth2\Client\Provider\Exception\IdentityProviderException;
@@ -22,28 +24,64 @@ class OAuthController extends ControllerBase {
   private $oauthClient;
 
   /**
+   * @var \Drupal\Core\Session\SessionManagerInterface
+   */
+  private $sessionManager;
+
+  /**
    * @var \Drupal\user\PrivateTempStore
    */
   private $tempStore;
 
   public function __construct(
     OAuthClient $oauthClient,
+    SessionManagerInterface $sessionManager,
     PrivateTempStore $tempStore
   ) {
     $this->oauthClient = $oauthClient;
+    $this->sessionManager = $sessionManager;
     $this->tempStore = $tempStore;
   }
 
-  public function beginAuthorization() {
+  public function beginAuthorization(Request $request) {
+    // Begin session for anonymous users, so private temp store can be used.
+    if ($this->currentUser()->isAnonymous() && !isset($_SESSION['session_started'])) {
+      $_SESSION['session_started'] = true;
+      $this->sessionManager->start();
+    }
+
     // Store state to prevent CSRF attacks later.
-    $this->tempStore->set('oauth.state', $this->oauthClient->getState());
+    $state = Crypt::randomBytesBase64(16);
+    $this->tempStore->set('oauth.state', $state);
+
     // Redirect user to authorization endpoint.
-    $this->oauthClient->authorize([
+    return $this->oauthClient->authorize([
       'scope' => [
         'read',
         'write',
       ],
-    ]);
+      'state' => $state,
+    ], [$this, 'getAuthorizationRedirect']);
+  }
+
+  /**
+   * Redirect callback function for the Zendesk OAuth client library.
+   *
+   * We use this function to return a redirect response instead of using the
+   * library's built in redirect function so Drupal has the opportunity to
+   * clean up gracefully.
+   *
+   * The main problem with not using a Drupal redirect is the session is not
+   * saved automatically, and so the 'state' parameter is lost & CSRF checks
+   * always fail.
+   *
+   * @param string $url
+   * @param \Stevenmaguire\OAuth2\Client\Provider\Zendesk $client
+   *
+   * @return \Drupal\Core\Routing\TrustedRedirectResponse
+   */
+  public function getAuthorizationRedirect(string $url, OAuthClient $client) {
+    return new TrustedRedirectResponse($url);
   }
 
   public function redirectEndpoint(Request $request) {
@@ -52,7 +90,7 @@ class OAuthController extends ControllerBase {
     }
 
     // Check state to prevent CSRF attacks.
-    if (!$request->query->has('state') || ($request->query->get('state') !== $this->tempStore->get('oauth.state'))) {
+    if (!($request->query->has('state') && ($request->query->get('state') === $this->tempStore->get('oauth.state')))) {
       $this->tempStore->delete('oauth.state');
       throw new AccessDeniedHttpException("Invalid state passed for OAuth authorization");
     }
@@ -75,15 +113,19 @@ class OAuthController extends ControllerBase {
       $this->login($user);
     } catch (IdentityProviderException $e) {
       // @todo Handle failure retrieving OAuth token/resource owner.
+      // Redirect to journey start with error message?
     }
   }
 
   public static function create(ContainerInterface $container) {
     /** @var \Stevenmaguire\OAuth2\Client\Provider\Zendesk $oauthClient */
     $oauthClient = $container->get('zendesk_connect.oauth_client');
+    /** @var \Drupal\Core\Session\SessionManagerInterface $sessionManager */
+    $sessionManager = $container->get('session_manager');
     /** @var \Drupal\user\PrivateTempStore $tempStore */
     $tempStore = $container->get('user.private_tempstore')->get('zendesk_connect');
-    return new static($oauthClient, $tempStore);
+
+    return new static($oauthClient, $sessionManager, $tempStore);
   }
 
   private function registerFromResourceOwner(ZendeskResourceOwner $resourceOwner): User {
